JAVA-3150: code review changes
