+++ driver-core/src/main/com/mongodb/management/JMXConnectionPoolListener.java
@@ -22,13 +22,11 @@ import com.mongodb.event.ConnectionCheckedInEvent;
 import com.mongodb.event.ConnectionCheckedOutEvent;
 import com.mongodb.event.ConnectionClosedEvent;
 import com.mongodb.event.ConnectionCreatedEvent;
-import com.mongodb.event.ConnectionPoolClearedEvent;
 import com.mongodb.event.ConnectionPoolClosedEvent;
 import com.mongodb.event.ConnectionPoolCreatedEvent;
 import com.mongodb.event.ConnectionPoolListener;
 import com.mongodb.event.ConnectionPoolWaitQueueEnteredEvent;
 import com.mongodb.event.ConnectionPoolWaitQueueExitedEvent;
-import com.mongodb.event.ConnectionReadyEvent;
 
 import javax.management.ObjectName;
 import java.util.List;
@@ -52,14 +50,6 @@ public class JMXConnectionPoolListener implements ConnectionPoolListener {
         ConnectionPoolStatistics statistics = new ConnectionPoolStatistics(event);
         map.put(event.getServerId(), statistics);
         MBeanServerFactory.getMBeanServer().registerMBean(statistics, getMBeanObjectName(event.getServerId()));
-    }
-
-    @Override
-    public void connectionPoolCleared(final ConnectionPoolClearedEvent event) {
-        ConnectionPoolStatistics statistics = getStatistics(event.getServerId());
-        if (statistics != null) {
-            statistics.connectionPoolCleared(event);
-        }
     }
 
     @Override
@@ -105,14 +95,6 @@ public class JMXConnectionPoolListener implements ConnectionPoolListener {
         ConnectionPoolStatistics statistics = getStatistics(event.getConnectionId());
         if (statistics != null) {
             statistics.connectionCreated(event);
-        }
-    }
-
-    @Override
-    public void connectionReady(final ConnectionReadyEvent event) {
-        ConnectionPoolStatistics statistics = getStatistics(event.getConnectionId());
-        if (statistics != null) {
-            statistics.connectionReady(event);
         }
     }
