+++ driver-core/src/test/functional/com/mongodb/internal/connection/DefaultConnectionPoolTest.java
@@ -27,7 +27,6 @@ import org.junit.Test;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static junit.framework.TestCase.assertNotNull;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -57,11 +56,11 @@ public class DefaultConnectionPoolTest {
     public void shouldThrowOnTimeout() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID, connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maxWaitTime(50, MILLISECONDS)
-                                                                   .build());
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maxWaitTime(50, MILLISECONDS)
+                        .build());
         provider.start();
 
         provider.get();
@@ -80,11 +79,11 @@ public class DefaultConnectionPoolTest {
     public void shouldThrowOnWaitQueueFull() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID, connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maxWaitTime(500, MILLISECONDS)
-                                                                   .build());
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maxWaitTime(500, MILLISECONDS)
+                        .build());
         provider.start();
 
         provider.get();
@@ -106,12 +105,12 @@ public class DefaultConnectionPoolTest {
     public void shouldExpireConnectionAfterMaxLifeTime() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID, connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maintenanceInitialDelay(5, MINUTES)
-                                                                   .maxConnectionLifeTime(50, MILLISECONDS)
-                                                                   .build());
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maintenanceInitialDelay(5, MINUTES)
+                        .maxConnectionLifeTime(50, MILLISECONDS)
+                        .build());
         provider.start();
 
         // when
@@ -128,11 +127,11 @@ public class DefaultConnectionPoolTest {
     public void shouldExpireConnectionAfterLifeTimeOnClose() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID,
-                                             connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maxConnectionLifeTime(20, MILLISECONDS).build());
+                connectionFactory,
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maxConnectionLifeTime(20, MILLISECONDS).build());
         provider.start();
 
         // when
@@ -148,12 +147,12 @@ public class DefaultConnectionPoolTest {
     public void shouldExpireConnectionAfterMaxIdleTime() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID,
-                                             connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maintenanceInitialDelay(5, MINUTES)
-                                                                   .maxConnectionIdleTime(50, MILLISECONDS).build());
+                connectionFactory,
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maintenanceInitialDelay(5, MINUTES)
+                        .maxConnectionIdleTime(50, MILLISECONDS).build());
         provider.start();
 
         // when
@@ -170,12 +169,12 @@ public class DefaultConnectionPoolTest {
     public void shouldCloseConnectionAfterExpiration() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID,
-                                             connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maintenanceInitialDelay(5, MINUTES)
-                                                                   .maxConnectionLifeTime(20, MILLISECONDS).build());
+                connectionFactory,
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maintenanceInitialDelay(5, MINUTES)
+                        .maxConnectionLifeTime(20, MILLISECONDS).build());
         provider.start();
 
         // when
@@ -192,12 +191,12 @@ public class DefaultConnectionPoolTest {
     public void shouldCreateNewConnectionAfterExpiration() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID,
-                                             connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maintenanceInitialDelay(5, MINUTES)
-                                                                   .maxConnectionLifeTime(20, MILLISECONDS).build());
+                connectionFactory,
+                ConnectionPoolSettings.builder()
+                        .maxSize(1)
+                        .maxWaitQueueSize(1)
+                        .maintenanceInitialDelay(5, MINUTES)
+                        .maxConnectionLifeTime(20, MILLISECONDS).build());
         provider.start();
 
         // when
@@ -215,14 +214,15 @@ public class DefaultConnectionPoolTest {
     public void shouldPruneAfterMaintenanceTaskRuns() throws InterruptedException {
         // given
         provider = new DefaultConnectionPool(SERVER_ID,
-                                             connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(10)
-                                                                   .maxConnectionLifeTime(1, MILLISECONDS)
-                                                                   .maintenanceInitialDelay(5, MINUTES)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .build());
-        provider.get().close();                                                  provider.start();
+                connectionFactory,
+                ConnectionPoolSettings.builder()
+                        .maxSize(10)
+                        .maxConnectionLifeTime(1, MILLISECONDS)
+                        .maintenanceInitialDelay(5, MINUTES)
+                        .maxWaitQueueSize(1)
+                        .build());
+        provider.get().close();
+        provider.start();
 
 
         // when
@@ -231,49 +231,5 @@ public class DefaultConnectionPoolTest {
 
         // then
         assertTrue(connectionFactory.getCreatedConnections().get(0).isClosed());
-    }
-
-    @Test
-    public void shouldNotCallWaitQueueExitedIfWaitQueueEnteredWasNotCalled() throws InterruptedException {
-        // given
-        QueueEventsConnectionPoolListener listener = new QueueEventsConnectionPoolListener();
-
-        // when
-        provider = new DefaultConnectionPool(SERVER_ID, connectionFactory,
-                                             ConnectionPoolSettings.builder()
-                                                                   .maxSize(1)
-                                                                   .maxWaitQueueSize(1)
-                                                                   .maxWaitTime(5, SECONDS)
-                                                                   .addConnectionPoolListener(listener)
-                                                                   .build());
-        provider.start();
-
-        // then
-        assertEquals(0, listener.getWaitQueueSize());
-
-        // when
-        InternalConnection connection = provider.get();
-        TimeoutTrackingConnectionGetter timeoutTrackingConnectionGetter = new TimeoutTrackingConnectionGetter(provider);
-        new Thread(timeoutTrackingConnectionGetter).start();
-        Thread.sleep(100);
-
-        // then
-        try {
-            provider.get();
-            fail();
-        } catch (MongoWaitQueueFullException e) {
-            assertEquals(0, listener.getWaitQueueSize());
-        }
-
-        // when
-        connection.close();
-        timeoutTrackingConnectionGetter.getLatch().await();
-        connection = provider.get();
-
-        // then
-        assertEquals(0, listener.getWaitQueueSize());
-
-        // cleanup
-        connection.close();
     }
 }
