--- driver-core/src/test/functional/com/mongodb/internal/connection/QueueEventsConnectionPoolListener.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2008-present MongoDB, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.mongodb.internal.connection;
-
-import com.mongodb.event.ConnectionPoolListener;
-import com.mongodb.event.ConnectionPoolWaitQueueEnteredEvent;
-import com.mongodb.event.ConnectionPoolWaitQueueExitedEvent;
-
-import java.util.concurrent.atomic.AtomicInteger;
-
-class QueueEventsConnectionPoolListener implements ConnectionPoolListener {
-    private final AtomicInteger waitQueueSize = new AtomicInteger();
-
-    @Override
-    public void waitQueueEntered(final ConnectionPoolWaitQueueEnteredEvent event) {
-        waitQueueSize.incrementAndGet();
-    }
-
-    @Override
-    public void waitQueueExited(final ConnectionPoolWaitQueueExitedEvent event) {
-        waitQueueSize.decrementAndGet();
-    }
-
-    int getWaitQueueSize() {
-        return waitQueueSize.get();
-    }
-}
